ISPN-10549 GlobalMarshaller's persistence marshaller should not use ExternalizerAwareSerializationMarshaller
By default the PersistenceMarshaller utilises the
ExternalizerAwareSerializationMarshaller to allow user configured
externalizers to be respected when persisting objects. However, in the
context of the GlobalMarshaller, which delegates to the
PersistenceMarshaller, this does not make sense as the GlobalMarshaller
already handles externalizers. Therefore, in order to prevent
unnecessary hash lookups, the PersistenceMarshaller used
by the GlobalMarshaller should delegate directly to the
JavaSerializatioMarshaller.